survey: cleanup survey routes
fixes #851

Split admin and participant routes for the survey access routes:
participants can only create and activate a survey for their own
participants. Admin app users do not have a own survey but need a UUID
to activate one.

Move the route setup from evolution-legacy to evolution-backend.
